% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RootGame.R
\name{RootGame}
\alias{RootGame}
\title{Generate the setup for a new game of Root}
\description{
Generate the setup for a new game of Root

Generate the setup for a new game of Root
}
\details{
Useful utility for getting a new game of Root started using an Advanced
setup draft. Tools are provided for choosing random maps, decks, and factions
as well as randomizing the clearings for any map (though Autumn requires a
bit of a hack).
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{map}}{The map of the game}

\item{\code{deck}}{The deck of the game}

\item{\code{factions}}{The factions available for a draft}

\item{\code{available_factions}}{The factions available to be picked}

\item{\code{picked_factions}}{The factions already picked in a draft}

\item{\code{players}}{The \emph{number} of players in a game}

\item{\code{vagabonds}}{The vagabonds available for a draft}

\item{\code{roster}}{A vector with the names of the players}

\item{\code{draft_order}}{A vector with the names of the players in draft order
(first element gets first draft)}

\item{\code{draft_order}}{A vector with the names of the players in play order
(first element gets first turn). This is the reverse of \code{draft_order}}

\item{\code{verbose}}{A boolean to control console output (defaults to \code{TRUE}
which allows for an interactive feel for the draft)}

\item{\code{verbose}}{A vector of clearing suits (first initial abbeviations to
aid with console output)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RootGame$new()}}
\item \href{#method-set_player_order}{\code{RootGame$set_player_order()}}
\item \href{#method-deal_factions}{\code{RootGame$deal_factions()}}
\item \href{#method-pick}{\code{RootGame$pick()}}
\item \href{#method-random_map}{\code{RootGame$random_map()}}
\item \href{#method-random_deck}{\code{RootGame$random_deck()}}
\item \href{#method-set_clearings}{\code{RootGame$set_clearings()}}
\item \href{#method-show_clearings}{\code{RootGame$show_clearings()}}
\item \href{#method-start_new_game}{\code{RootGame$start_new_game()}}
\item \href{#method-clone}{\code{RootGame$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new Root game object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RootGame$new(
  map = NA,
  deck = NA,
  to_exclude = NA,
  players = 4,
  vagabonds_to_exclude = NA,
  roster = NA,
  verbose = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{map}}{Map of the game}

\item{\code{deck}}{Deck of the game}

\item{\code{to_exclude}}{Factions to exclude from the draft}

\item{\code{players}}{Number of player (overwritten by the length of \code{roster})}

\item{\code{vagabonds_to_exclude}}{Vagabond characters to exclude from the draft}

\item{\code{roster}}{The names of the players in the game}

\item{\code{verbose}}{Boolean to control console output}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{RootGame} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_player_order"></a>}}
\if{latex}{\out{\hypertarget{method-set_player_order}{}}}
\subsection{Method \code{set_player_order()}}{
Sets the draft and turn order of the players (\code{roster} must be a valid
vector of players)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RootGame$set_player_order()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-deal_factions"></a>}}
\if{latex}{\out{\hypertarget{method-deal_factions}{}}}
\subsection{Method \code{deal_factions()}}{
Creates a valid Advanced Setup draft of factions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RootGame$deal_factions()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pick"></a>}}
\if{latex}{\out{\hypertarget{method-pick}{}}}
\subsection{Method \code{pick()}}{
An interactive function for going through the draft; accepts either the
fully spelled out faction name or the index of the faction. If the last
faction is white, it will not allow it to be chosen until a red faction
has been chosen.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RootGame$pick(faction)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-random_map"></a>}}
\if{latex}{\out{\hypertarget{method-random_map}{}}}
\subsection{Method \code{random_map()}}{
Simple utility for chosing a random map
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RootGame$random_map()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-random_deck"></a>}}
\if{latex}{\out{\hypertarget{method-random_deck}{}}}
\subsection{Method \code{random_deck()}}{
Simple utility for chosing a random deck
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RootGame$random_deck()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_clearings"></a>}}
\if{latex}{\out{\hypertarget{method-set_clearings}{}}}
\subsection{Method \code{set_clearings()}}{
Simple utility for setting the clearings of a map
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RootGame$set_clearings(force = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-show_clearings"></a>}}
\if{latex}{\out{\hypertarget{method-show_clearings}{}}}
\subsection{Method \code{show_clearings()}}{
Utility for printing the clearings to the console in a reasonable way
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RootGame$show_clearings()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-start_new_game"></a>}}
\if{latex}{\out{\hypertarget{method-start_new_game}{}}}
\subsection{Method \code{start_new_game()}}{
Wrapper function that will choose a map & deck, set clearings, and start
up a draft.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RootGame$start_new_game(map = "random", deck = "random", force = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RootGame$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
